+++ chemreac/serialization.py
@@ -5,7 +5,7 @@ import json
 import numpy as np
 
 from . import ReactionDiffusion
-from .core import get_unit
+from .core import get_unit, g_units, k_units
 from .units import (
     unit_registry_to_human_readable,
     unit_registry_from_human_readable
@@ -38,13 +38,13 @@ def dump(rd, dest):
         'n': rd.n,
         'stoich_active': rd.stoich_active,
         'stoich_prod': rd.stoich_prod,
-        'k': list(rd._k),
-        'D': list(rd._D),
-        'mobility': list(rd._mobility),
+        'k': list(rd.k),
+        'D': list(rd.D),
+        'mobility': list(rd.mobility),
         # 'x': rd.x.tolist(),
         'stoich_inact': rd.stoich_inact,
         'units': unit_registry_to_human_readable(rd.unit_registry),
-        'g_values': list(rd._g_values),
+        'g_values': list(rd.g_values),
         'g_value_parents': rd.g_value_parents,
         # 'fields': fields
     }
@@ -84,10 +84,10 @@ def load(source, RD=None, **kwargs):
     if 'g_values' in data:
         data['g_values'] = [elem*g_unit for elem, g_unit in
                             zip(np.array(data['g_values']),
-                                RD.g_units(unit_registry,
-                                           data['g_value_parents']))]
+                                g_units(unit_registry,
+                                        data['g_value_parents']))]
     reaction_orders = map(len, data['stoich_active'])
-    kunits = RD.k_units(unit_registry, reaction_orders)
+    kunits = k_units(unit_registry, reaction_orders)
     data['k'] = [kval*kunit for kval, kunit in zip(data['k'], kunits)]
     data.update(kwargs)
     extra_data = {}