+++ chemreac/util/tests/test_analysis.py
@@ -13,8 +13,7 @@ def test_solver_linear_error():
 @pytest.mark.parametrize("atol", (1.0, [1.0]))
 def test_solver_linear_error_from_integration(atol):
     class Dummy:
-        def with_units(self, value, key):
-            return value
+        unit_registry = None
     integr = Dummy()
     integr.yout = np.array([1.0]).reshape((1, 1, 1))
     print(integr.yout[slice(None), 0, 0])